CMakeLists: only install headers when loader is being installed
There are scenarios where various components such as
Vulkan-LoaderAndValidationLayers, VulkanTools etc
are installed side by side on the same system. The
header deployment gets in to a conflicting situation
in such setups.
Now we conditionalize the deployment so that the package
building the loader will install the headers as well.

Signed-off-by: Awais Belal <awais_belal@mentor.com>